tutor register controller added
tutor register controller added with length, transaction-based sql query. The coee works with a postman post request to the relevant endpoint. Not yet tested in  browser. Need to also add registration for students as well as login and JWT implementation.
Also need to amend the FindTutors component as it no longer find tutors; this may be because database schema was altered by adding in a user roles table as well as an app-users table that will be used for auth logins going forward