Calendar component extra functionality
Added extra frotnend functionality: different styling for draft and confirmed events. Cancel bookings. Check for overlapping bookings and prevent it occuring. Added a panel that displays event info on clicking the event. and is removed when the event is clicked again. Added functions for formatting the date and times of events in to nicer strings for UI.