added personalised tutor links via parameterisation and standardsised errors on backlend in most controllers
On frontend - created parameterised /:tutorId links via react router to get individual tutor details after clicking on a tutor's profile card in the 'find a tutor' page. This was achieved by embedding the tutor profile components in a 'Link' component that linked to a new TutorProfilePage with the tutorId taken from the url params and used for  fetching the data and rendering it in a TutorCard.

On backend
refactored checkemails.sql query used in findUserByEmail() function in login controller so that it obtained all the possible user data which means there is only 1 DB query in the controller when previously there were 3 separate queries in the controller.

Began standardisation of error object format creation by creating new functions within new uti/httpError.mjs file.
Standardised errors first set up in the login controller in  authController.mjs file.
Thgen standardised errors in the getTutors and getTutorById controllers in the getTutorsControllers.mjs file