standardised error handling
standardised error handling in backend. Simplified it down from what was rpeviously attempted . the approach is simply to set a res.status() and then throw an error with simple message. and pass thsi to a global error handler. The global error handler then sens the status code and a json respons econtaining the error message.
Only the form validation errors remain unchanegd so that they can send extra detail if needsbe but, currently, the  frontend handles that before submission to backend is even allowed.