standardised  form dropdown behaviour
Brought the more complete beahviour from the Tutor register form to the student register form.  This refers to the instrument/city dropdown behaviour that ensures the dropdown disappears when exact match is found- the bug was that e.g. when clicking piano, electric piano and jazz piano would still show.

 Removed some unused compnent files from frontend