frontend appointment booking
created BookingCalendar component based  off of the CalendarPRacticePage (no longer used). The new component is rendered as part of the TutorProfile page and context is used to provide the component the student details. Tutor info is fetched by the TutorProfile page so is just passed in as a prop as normal without using context.

Added some UX tewaks to have calendar component only viewable if a student is logged in and even then diplay is toggled by button.

Event details panel was amaneded to show more details.

Login auth handler was changed ever-so-slightly  so that both tutor and user have .first_name and .last_name for consistent proeprty access  for redering.

Amended city-dropdown logic in the student registration form to be like the city-dropdown in the tutor regiatration form to eloimate potential bug where dropdown remained if matrching cities remained that contained the selected city e.g. slected 'York' but 'New York ' remains visible.