refactoring to begin backed appointment booking
the TutorPRofilePage and the BookingCalendar component have been refactored to begin creating frontend to backend appointment booking.
Main changes are that the TutorProfilPage now fetches the  tutor and student appointment data and then transfroms it for the Cakendar and this is passed as a prop to the calendar. Also, the handleCOnformBooking and handleCancelBooking functions hae been lifted up to parent component as they will perform backend fetch calls to amend appointment table in DB,

In summary, the TuotProfilePage will handle most of the appointment state and  frontend validation checks and the BookingCalendar component will primarily handle UI for rendering the appointments and draft appointments.

SOme of the validity checking has been parcelled of in to a new, inididvual function that will be used before attemtping backend booking. A  'master function' that calls both the booking-checking function and the handleConfirmBooking function has been written in the Claendar component.

The handleconfirmappointment function itself still needs to be written to fetch at backend.

backend fucntionality needs to be writtend to verify checks at backend before commiting appointments to DB.