appointment booking functionality ongoing implementation
slightly refactored the  getBookingsByTutorId and getBookingsByStudentId handlers in bookingController.mjs  and added a new, basic handler for making biikings from booking info received from frontend. This handler is called makeBooking

in the getTutors controller file, I  added a new handler 'getTutorById' that is used to  get individual tutor data for the new FocusedTutorCard component. The handler fetches tutor data by t:tutorId param and runs 5 separate sql queries in order to build a  bespoke tutor object where proerties with multiole entries e.g. 'teaching format' are attached as arrays. This object is sent  to frontend which then reders releavnt details in the new FocusedTutoCard component and the options are rendered in a clcickabel button that allows student to select these to determine a particular aspect of the draft lesson e.g. click 'imn-person' for in person lesson etc.

Amended the getBookingsByStudentId.sql and getBookingsByTutor.sql queries to return more info

Created a query for inserting draft appointment info from frontend in to backenbd and DB

Enitely removed the unusued getTutorById.sql file as there is a more specific one actually in use

As per above, I created new, small sql queries for getting tutor info to create bespoke tutor object for use in the FocusedTutorCard compknent.

Created the FocusedTutorCard component to replace the old card. This new component fetches and renders indiiviudal tutor data in a specialised way to allow easier booking.

Updated the router in BookingRoutes to have route and handler for  making bookings

Made some siginificant changes to the BookingsCalendar component in order to allign it with the new frontend-backend data transfer for making appointments. examole, createrd a new mutuator for draftEvent state.
changed the draftBooking object to use only ids with its proeprties rather than text. These ids correspond to text info in the database.
Changed the logic test for determining if a draft event can be rendered in Calendar compoinent and whtehr a draft booking can be confirmed i.e. the isDraftValid variable was rpalced by a candisplayDraft checker and a canConfirmDraft checker.

For clarity and remvoal of potential variable duplciation with new code, the booking was changed from draftEvent to draftBooking  and draftEvent vsriable name was just used as a genrric paramater name  in definitions of some functions in the file.

State was lifted from Calendar and FocusedTutorCard in to the parent component i.e. TutorPRofilePage. E.g. handleConffirmEvent() and handleCancelEvent() as well as draftBooking and updateDraftBooking. These were then passed as props to  FocusedTutorCard and BookingCalendar as required.

The BookingCalendar's handleDraftBooking() function and preBookingChecks() functions were amended in line with new prop data structure. In draftBookings and anywhere it is accessed by finctions, the 'start' and 'end' properties wer changed to 'booking_start_time' and 'booking_end_time' in order to align with the relevant column names  in database.

BookingCalendar's clearSelection() function updated accrodingly. The isSameEvent() function was updated to deliberately have looser equality checks to proeprly handle nulls and undefined to avpid incorrectly assuming 2 eevents with indefined could be considered the same.

in thw TutorProfilePage, I added a parsetutorId() function to check the :tutorId URL param to prevent fetches in the relavnt useEffect from running in the tutrId was not a positive integer and refactored any functions from the tutorId to parsedTutorId.

Amended the bookingShaper() helper function so it hsaped events to have booking_start_time and booking_end_time proeprties. UPdated the Calendar component's start and end props to reference booking_start_time and booking_end_time to ensure events could still be displayed