appointment booking functionality further improvements
Amended theTutorProfile page by adding in a useEffect that will amend the draftBookings student and tutor ids if either of them change. This is not strictly required right but could be needed if I implement code to keep users logged in after browser refresh.

In the handleConfirmBooking() function, I added another saftety check for draftBookings where, as the draft booking details are being sent to backend, check for the current tutor and student ids incase they became stale. Again, not strictly necessary right now but could be in future.
In this function I also waited the appointment booking confirmed data to come from the DB and then added this booking detail in to the tutorBookings array so that it newly booked appointments would show in the Calendar straight away.  Also added some basic error handling and the function now throws error outwards to be handled by  the confirmDraftBooking() function in BookingCalendar that calss handleConfiemBooking()

I added the keyword 'distinct' to the tutorTeachingTypes.sql query used in the getTutorById route controller in order to reduce the duplication of teaching types in UI

Amended the backend's global erorr handler to take res.status and err.stus in to account

Added basic validation check to the makeBooking route handler.

Re-standardised responses and error responses across all handlers.

added detail to  login controller in authComtroller file to handle admin user

added more error checking in the makeBooking handler

Global errro handler was refined in  order to take in to account the  res.status codes set in handlers rather than just lebling most things as 500 status